hotfix: empty adapt fix
